Move Dashboard into own component with own nav
