+++ src/app/dashboards/dashboards.component.html
@@ -1,4 +1,4 @@
 <div class="main-container">
     <header class="header header-7" app-header></header>  
-    <p>Welcome to the Admin Page</p>
+    <progress-dashboard class="progress-dashboard"></progress-dashboard>
 </div>
