+++ src/app/dashboards/dashboards.component.ts
@@ -1,0 +1,9 @@
+import { Component} from '@angular/core';
+
+@Component({
+  selector: 'app-dashboards',
+  templateUrl: './dashboards.component.html',
+  styleUrls: ['./dashboards.component.scss']
+})
+export class DashboardsComponent {
+}
