+++ src/app/dashboards/progress-dashboard/progress-dashboard.component.html
@@ -1,0 +1,134 @@
+<div class="content-container">
+    <nav class="sidenav">
+        <clr-tabs>
+            <clr-tab>
+              <button clrTabLink>Sessions</button>
+              <ng-template [(clrIfActive)]="sessionDashboardActive">
+              <clr-tab-content>
+
+                <section class="nav-group collapsible">
+                    <input id="activeEventsToggle" type="checkbox" />
+                    <label for="activeEventsToggle">Active Events</label>
+                    <ul class="nav-list">
+                        <li *ngFor="let se of activeEvents" (click)="setScheduledEvent(se)" [ngClass]="{'selected' : selectedEvent == se}">
+                            <div class="clr-row">
+                                <div class="nav-event-name" [class.short]="!!se.activeSessions">{{se.event_name}}</div> 
+                                <div *ngIf="se.activeSessions != undefined && se.activeSessions != 0" class="active-sessions-bubble ">{{se.activeSessions}}</div>
+                            </div>
+                        </li>
+                    </ul>
+                    <input id="finishedEventsToggle" type="checkbox" checked="!checked" />
+                    <label for="finishedEventsToggle">Finished Events</label>
+                    <ul class="nav-list">
+                        <li *ngFor="let se of finishedEvents" (click)="setScheduledEvent(se)" [ngClass]="{'selected' : selectedEvent == se}">
+                            <div class="clr-row">
+                                <div class="nav-event-name" [class.short]="!!se.activeSessions">{{se.event_name}}</div> 
+                                <div *ngIf="se.activeSessions != undefined && se.activeSessions != 0" class="active-sessions-bubble ">{{se.activeSessions}}</div>
+                            </div>
+                        </li>
+                    </ul>
+                </section>
+
+              </clr-tab-content>
+            </ng-template>
+            </clr-tab>
+            
+            <clr-tab>
+                <button clrTabLink>VMs</button>
+                <ng-template [(clrIfActive)]="vmDashboardActive">
+                <clr-tab-content *clrIfActive>
+  
+                  <section class="nav-group collapsible">
+                      <input id="activeEventsToggle" type="checkbox" />
+                      <label for="activeEventsToggle">Active Events</label>
+                      <ul class="nav-list">
+                          <li *ngFor="let se of activeEvents" (click)="setScheduledEvent(se)" [ngClass]="{'selected' : selectedEvent == se}">
+                              <div class="clr-row">
+                                  <div class="nav-event-name" [class.short]="!!se.activeSessions">{{se.event_name}}</div> 
+                                  <div *ngIf="se.activeSessions != undefined && se.activeSessions != 0" class="active-sessions-bubble ">{{se.activeSessions}}</div>
+                              </div>
+                          </li>
+                      </ul>                      
+                  </section>
+  
+                </clr-tab-content>
+            </ng-template>
+              </clr-tab>
+            
+        </clr-tabs>
+    
+</nav>
+<div *ngIf="sessionDashboardActive" class="content-area">
+    <h3>Active Sessions for {{selectedEvent?.event_name}}</h3>
+    <span class="additional-info">AC: <span class="access-code">{{selectedEvent?.access_code}}</span> &#8226; Until {{selectedEvent?.end_time | date:'MMM d, y, HH:mm'}} <span>&#8226; Created by {{ selectedEvent?.creatorEmail }}</span></span>
+    <div class="clr-row clr-align-items-center clr-justify-content-between">
+    <div class="card clr-col-9">
+        <div class="clr-row clr-align-items-center">
+            <div class="clr-col">                  
+                <clr-toggle-container id="includeFinishedToggle">
+                    <clr-toggle-wrapper>
+                    <input type="checkbox" clrToggle name="includeFinished" [disabled] = "selectedEvent?.finished" [(ngModel)]="includeFinished" (change)="refresh()">
+                    <label>Include finished</label>
+                    </clr-toggle-wrapper>
+                    <clr-control-helper>Whether finished sessions should be included.</clr-control-helper>
+                </clr-toggle-container>                
+            </div>
+            <div class="clr-col ">     
+                <input id="userFilter" clrInput placeholder="filter users" name="userFilter" [(ngModel)]="userFilter" (ngModelChange)="filter()" [disabled] = "filteredProgress.length == 0 && userFilter.length == 0"/>
+                <span class="clr-subtext">Filter by users email</span> 
+            </div>
+            <div class="clr-col">
+                <clr-dropdown >
+                    <button class="dropdown-toggle btn btn-link" clrDropdownTrigger [disabled] = "scenarioList.size < 2">
+                    <span *ngIf="this.scenarioFilterList.size == 0">Filter scenario(s)</span>
+                    <span *ngIf="this.scenarioFilterList.size == 1">1 Scenario selected</span>
+                    <span *ngIf="this.scenarioFilterList.size > 1">{{this.scenarioFilterList.size}} Scenarios selected</span>
+                    <clr-icon shape="caret down"></clr-icon>
+                    </button>
+                    
+                    <clr-dropdown-menu clrPosition="bottom-right" *clrIfOpen>                          
+                        <clr-row *ngFor="let sc of scenarioList" clrDropdownItem (click)="filterScenario(sc)" [ngClass]="{'selected' : this.scenarioFilterList.has(sc)}">
+                            {{sc}}
+                            <span class="filter-selected" >
+                                <clr-icon *ngIf="this.scenarioFilterList.has(sc)" shape="check" ></clr-icon>
+                            </span>
+                        </clr-row>                            
+                    </clr-dropdown-menu>
+                </clr-dropdown>
+            </div>
+            <div class="clr-col">
+                <p>Showing {{filteredProgress.length}} of {{currentProgress.length}} Sessions</p>
+                <button class="btn btn-icon btn-warning" (click)="removeFilter()"><clr-icon shape="trash"></clr-icon> Clear Filter</button> 
+            </div>            
+        </div>            
+    </div>        
+    <div class="clr-col buttonCol" >
+        <button class="btn btn-link userlistIcon" [disabled] = "users?.length < 1" (click)="openUserList()">
+            <clr-icon  shape="users" class="is-solid" size="48"></clr-icon>{{users?.length}}
+        </button>
+    </div>                        
+    <div class="clr-col"> 
+        <interval-timer (intervalElapsed)="refresh()"></interval-timer>
+    </div>      
+</div>
+    <div class="clr-row">
+        <ng-container *ngIf="currentProgress?.length > 0; else no_sessions">
+            <div class="clr-col-12 clr-col-sm-6 clr-col-md-4 clr-col-lg-3" *ngFor="let p of filteredProgress">
+                <progress-card [progress] = "p" [pause] = "pause" (nameClickedEvent)="filterName($event)"></progress-card>
+            </div>
+        </ng-container>
+    </div>
+    <ng-template #no_sessions>
+        <div class="clr-col-12">
+          <p>
+            No sessions found.
+          </p>
+        </div>
+      </ng-template>                     
+</div>
+<div *ngIf="vmDashboardActive" class="content-area">
+    <p> This will be a new Dashboard showing cool info about VMs </p>
+</div>
+</div>
+
+<event-user-list #userList [users]="users" [progress]="currentProgress" (userSelected)="filterName($event.email)"></event-user-list>
