+++ src/app/dashboards/progress-dashboard/progress-dashboard.component.scss
@@ -1,0 +1,122 @@
+.content-container {
+    height: 100%;
+}
+
+span.bignum {
+    font-size: 6em;
+}
+
+span.access-code{
+  user-select: all;
+  cursor: pointer;
+}
+
+.nav-group{
+  overflow-y: auto;
+  overflow-x: hidden;
+}
+
+.nav-event-name {
+  width: 100%;
+  text-overflow: ellipsis;
+  overflow: hidden;
+  white-space: nowrap;
+  padding: 0 0 0 10px;
+}
+.nav-event-name.short {
+  width: 80%;
+}
+
+ul > li.selected{
+    background-color: #eaeaea;
+    font-weight: bold;
+    color: black;
+}
+
+ul.nav-list > li {
+    padding: 10px;
+}
+
+ul.nav-list > li:first-child {
+    margin-top: 5px;
+}
+
+li .active-sessions-bubble{
+    margin-right: 10px;
+    background-color: white;
+    padding: 0 10px;
+    border-radius: 100%;
+}
+
+.filter-selected{
+    float: right;
+    margin-left: 20px;
+    width: 10px;
+    height: 10px;
+}
+
+#includeFinishedToggle {
+    margin-bottom: 1.2rem;
+}
+
+clr-row.selected{
+    background-color: #f0f0f0;
+}
+
+
+.buttonCol{
+  text-align: right;
+  margin-top: 1.2rem;
+}
+
+.countdown-container{
+  float:right;
+}
+
+#countdown {    
+    position: relative;
+    height: 40px;
+    width: 40px;
+    text-align: center;
+    cursor: pointer;
+    margin: 2px auto;
+}
+  
+#countdown-number {
+    color: black;
+    display: inline-block;
+    line-height: 40px;
+  }
+  
+svg {
+    position: absolute;
+    top: 0;
+    right: 0;
+    width: 40px;
+    height: 40px;
+    transform: rotateY(-180deg) rotateZ(-90deg);
+  }
+  
+svg circle {
+    stroke-dasharray: 113px;
+    stroke-dashoffset: 0px;
+    stroke-linecap: round;
+    stroke-width: 2px;
+    stroke: black;
+    fill: none;
+    animation: countdown 10s linear infinite forwards;    
+  }
+  
+@keyframes countdown {
+    from {
+      stroke-dashoffset: 0px;
+      stroke: green;
+    }
+    40% {
+       stroke: orange; 
+    }
+    to {
+      stroke-dashoffset: 113px;
+      stroke: red;
+    }
+  }
