+++ src/app/header/header.component.html
@@ -7,6 +7,7 @@
 
 <div class="header-nav">
     <a href="javascript://" class="nav-link nav-text" [routerLink]="['/home']" [routerLinkActive]="'active'">Home</a>
+    <a href="javascript://" class="nav-link nav-text" [routerLink]="['/dashboards']" [routerLinkActive]="'active'">Dashboards</a>
     <a href="javascript://" class="nav-link nav-text" [routerLink]="['/events']" [routerLinkActive]="'active'">Scheduled Events</a>
     <a href="javascript://" class="nav-link nav-text" [routerLink]="['/content']" [routerLinkActive]="'active'">Content Management</a>
     <a href="javascript://" class="nav-link nav-text" [routerLink]="['/users']" [routerLinkActive]="'active'">Users</a>
